+++ R/tm_g_gh_spaghettiplot.R
@@ -383,10 +383,12 @@ srv_g_spaghettiplot <- function(id,
 
   moduleServer(id, function(input, output, session) {
     # reused in all modules
-    anl_q <- constr_anl_q(
+    anl_q_output <- constr_anl_q(
       session, input, data, dataname,
       param_id = "xaxis_param", param_var = param_var, trt_group = input$trt_group, min_rows = 1
     )
+
+    anl_q <- anl_q_output()$value
 
     # update sliders for axes taking constraints into account
     yrange_slider <- toggle_slider_server("yrange_scale")
@@ -395,15 +397,34 @@ srv_g_spaghettiplot <- function(id,
 
     horizontal_line <- srv_arbitrary_lines("hline_arb")
 
+    iv_r <- reactive({
+      iv <- shinyvalidate::InputValidator$new()
+
+      iv$add_rule("xaxis_param", shinyvalidate::sv_required("Please select a biomarker"))
+      iv$add_rule("trt_group", shinyvalidate::sv_required("Please select a treatment variable"))
+      iv$add_rule("xaxis_var", shinyvalidate::sv_required("Please select an X-Axis variable"))
+      iv$add_rule("yaxis_var", shinyvalidate::sv_required("Please select a Y-Axis variable"))
+
+      iv$add_rule("facet_ncol", shinyvalidate::sv_required("Number of plots per row must be a positive integer"))
+      iv$add_rule("facet_ncol", shinyvalidate::sv_integer("Number of plots per row must be a positive integer"))
+      iv$add_rule(
+        "facet_ncol", shinyvalidate::sv_gt(0, message_fmt = "Number of plots per row must be a positive integer")
+      )
+
+      iv$add_validator(horizontal_line()$iv_r())
+      iv$add_validator(anl_q_output()$iv_r())
+      iv$enable()
+      iv
+    })
+
+
     plot_q <- reactive({
+      teal::validate_inputs(iv_r())
       req(anl_q())
       # nolint start
       ylim <- yrange_slider$state()$value
       facet_ncol <- input$facet_ncol
-      validate(need(
-        is.na(facet_ncol) || (as.numeric(facet_ncol) > 0 && as.numeric(facet_ncol) %% 1 == 0),
-        "Number of plots per row must be a positive integer"
-      ))
+
       rotate_xlab <- input$rotate_xlab
       hline_arb <- horizontal_line()$line_arb
       hline_arb_label <- horizontal_line()$line_arb_label
@@ -415,8 +436,6 @@ srv_g_spaghettiplot <- function(id,
       trt_group <- input$trt_group
 
       # Below inputs should trigger plot via updates of other reactive objects (i.e. anl_q()) and some inputs
-      validate(need(input$xaxis_var, "Please select an X-Axis Variable"))
-      validate(need(input$yaxis_var, "Please select a Y-Axis Variable"))
       param <- input$xaxis_param
       xaxis_var <- input$xaxis_var
       yaxis_var <- input$yaxis_var